scripts/clipper/requirements.txt
yt-dlp
google-generativeai
moviepy<2.0
python-dotenv
